add basic test for expanding and collapsing a node
It fails due to soem error thrown deeply inside of ember-animated:

```
TypeError: yield* (intermediate value)(intermediate value) is not iterable
```

Need to look into that one separately. The error is also thrown in the
dummy app. So it does not seem to be specific to test setup but may also
affect production.